Overhaul build system.
This commit changes the Makefile to build the compiler incrementaly.
It also moves tests from a dedicated script into the Makefile.

A CMakeLists.txt and compile.sh are added to help compile the compiler
on platforms without a good GNU Make.

-MD support was added.

The run_tests_with_mingw_and_wine.sh was added to make the process
of testing windows builds more streamlined.